boot: add clock configuration
This patch lays out how clock configuration should come about by
providing an example for STM32F302R8 and modifying the boot1 stage to
call it at an appropriate moment.

This is quite a crucial step as it now enables us to setup SysTick,
which will likely be used by the scheduler in the future.

Signed-off-by: Duszku <duszku511@gmail.com>