fault: Implement a basic fault handler
This patch adds a basic fault handler that disables IRQs, collects
register values and prints a panic message before hanging on an infinite
loop. This is by no means finished, as fault registers could be read too
in order to provide the user with more information, but it is a good way
to start.

Some ways of testing this include calling the UDF assembly instruction,
disabling FPU and accessing coprocessor system register or enabling trap
on division by zero and dividing by zero.

Signed-off-by: Duszku <duszku511@gmail.com>