spinlock: Implement spinlocks
This patch adds first synchronization primitive - the spinlock. The only
board that is currently supported only has one core, but I do plan on
supporting SMP in the future, and since next big thing on the list is
dynamic memory allocation, I wanted to add some locking mechanism to
allow for future-proof code.

Signed-off-by: Duszku <duszku511@gmail.com>