alloc: Add a basic memory manager
This patch focuses on implementing a very basic dynamic allocation
mechanism. For now it is a first-fit algorithm. It is meant to be
replaced with something better in the future.

The reason we take this approach is that dynamic memory management is a
necessary prerequisite for scheduler that supports dynamic creation and
removal of tasks. The philosophy of this project, since its a one-man
hobby, is to always start small and build up when needed (or when
inspiration strikes.)

Signed-off-by: Duszku <duszku511@gmail.com>