syscall: Implement syscall handling
This patch implements an SVCall IRQ handler coupled with some X-macro
magic that allows declaring and calling syscalls.

Adding a new syscall requires two actions:

  1. Implementing a handler declared using SYSCALL_VOID or
     SYSCALL_NONVOID macros.
  2. Adding an entry to SYSCALLS X macro in syscall.h

This will automatically generate a wrapper function sys_<name> that
calls the syscall, passes the arguments and, if the syscall returns a
value - relays that returned value to the caller.

Signed-off-by: Duszku <duszku511@gmail.com>