ctx: Move context-collection to a new module
So far the only module requiring ability to collect context and work
with stacked IRQ frames was the fault handler. This will change soon
with introduction of a scheduler. Therefore this patch sections off the
code related to assembling a full context into a separate 'context'
module, to provide an API that can be used by the scheduler in the
future.

Signed-off-by: Duszku <duszku511@gmail.com>