panic: Add a panic syscall
This patch extends panic beyond just a reaction to faults. By adding a
panic system call it is now possible to trigger a panic from software
upon detection of serious errors. Its first use can be seen in alloc
subsystem, for block list corruptions.

This feature is not yet complete. In order to implement a proper message
printing we first need snprintf. And that in turn requires a refactor of
log API.

Signed-off-by: Duszku <duszku511@gmail.com>