_pfmt: Add a shared print format processor
So far the only module that worked with formatted output was the log
API. This, however, will not be the case forever, as I intend to add
snprintf() soon. As such I've separated the code that processes format
specifiers and produces outputs from the code that decides where this
output goes. I have also added support for %d and %p format specifiers
together with better option support.

As a final addition I have secured log API with a spinlock to avoid
messy writes on multi-core systems.

Signed-off-by: Duszku <duszku511@gmail.com>